refactor(backend): separate mod manager, errors, and rename logic
Changes:
- Moved the mod manager implementation into `manager.rs`
- Moved ModError into `errors.rs`
- Moved mod renaming logic into `mods/rename.rs`
- Validate mod names before renaming (empty name, invalid chars, etc)
