+++ src-tauri/src/state.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc, Mutex};
 
-use crate::{BD2ModManager, config::BD2Config};
+use crate::{manager::BD2ModManager, config::BD2Config};
 
 pub struct BundledAssets(pub std::collections::HashSet<String>);
